// src/Frontend/ViteApp/src/api-generated/models/StoreTransactionReadResponse.ts
export interface StoreTransactionReadResponse {
    /**
     * 
     * @type {CashBoxesReadIdParameter}
     * @memberof StoreTransactionReadResponse
     */
    id: CashBoxesReadIdParameter;
    /**
     * 
     * @type {string}
     * @memberof StoreTransactionReadResponse
     */
    note: string | null;
    /**
     * 
     * @type {Date}
     * @memberof StoreTransactionReadResponse
     */
    startedAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof StoreTransactionReadResponse
     */
    cancelledAt: Date | null;
    /**
     * 
     * @type {UserListModel}
     * @memberof StoreTransactionReadResponse
     */
    startedBy: UserListModel;
    /**
     * 
     * @type {UserListModel}
     * @memberof StoreTransactionReadResponse
     */
    cancelledBy: UserListModel | null;
    /**
     * 
     * @type {TransactionReason}
     * @memberof StoreTransactionReadResponse
     */
    reason: TransactionReason;
    /**
     * 
     * @type {ContainerUpdateModelPipeId}
     * @memberof StoreTransactionReadResponse
     */
    saleTransactionId: ContainerUpdateModelPipeId | null;
    /**
     * 
     * @type {Array<StoreTransactionItemModel>}
     * @memberof StoreTransactionReadResponse
     */
    storeTransactionItems: Array<StoreTransactionItemModel>;
}